fix: add back missing properties for tinysnb's schema (#233)
Committed-by: bingqing.lbq from Dev container

<!--
Thanks for your contribution! please review
https://github.com/GraphScope/neug/blob/main/CONTRIBUTING.md before
opening an issue.
-->

## What do these changes do?

<!-- Please give a short brief about these changes. -->

As titled. This pr add such missing properties back into the tinysnb
schema.

However, PLEASE NOTE that there will be still differences in the schema
definition for Neug compared to the one used in Kuzu (and therefore in
the tests) due to varying supported data types.
For instance, Neug currently does not support types such as INT8, UINT8,
INT16, UINT16 and represents them as INT32/UINT32. Similarly, types like
INT128, STRUCT, UNION, and MAP are not supported and are stored as
STRING.

## Related issue number

<!-- Are there any issues opened that will be resolved by merging this
change? -->

Fixes #232