feat: Support open neug db with empty data dir (#298)
https://github.com/GraphScope/neug/issues/47

When data_dir is empty, Neug will create dir `neug_db_{timestamp}`in the
temporary directory as the working directory.

The temporary directory is specified by the environment variable
`NEUG_DB_TMP_DIR`; if this environment variable is empty, the `/tmp`
directory will be used as the temporary directory.

When using Neug db with an empty data_dir, the temporarily created
working directory will be deleted when the database is closed.