docs: Documentation not necessarily depends on package building (#476)
Previously, the documentation depends on building the python package
first, which is not expected.
Since we don't care much about the binding code's api, we will not
include the binding code's documentation in python reference, which
allows us to generate documentation right from the python code.

An environment variable `BUILD_DOC` is introduced to control whether to
report the import error.

Fix #454 
Fix #477
Fix #478