fix: Minor fix and turn on cyclic and arithmetic tests (#491)
For expressions built for `ProjectExpr`, `is_const_expr` is required to
tell the engine whether it holds `ConstExpr`.
If it is, then even if there is no data in the Context, the `constexpr`
will be called once to generate a single record.
If it is not, then the project will follow the normal projection logic.