refactor: Remove the dependency for proto-compiler for embed mode (#656)
Fix #652 

To resolve this issue, the generated `pb.h` and `pb.cc` files have been
added directly to the repository. This change eliminates the dependency
on a global protobuf compiler for file generation. Additionally, by
including Protobuf as a third-party library within the project, reliance
on system-installed Protobuf libraries is avoided in embedded mode.

### Important Notes

- The protobuf compiler is still required for the service mode. In
scenarios where the protobuf compiler is installed globally, but the
library is not, the project will utilize the submodule version of
Protobuf within the codebase. It's important to ensure that the version
of `protoc` used is 3.21.9 or later.
- The autogenration of pb.h and pb.cc are commented out.