fix: Rename kuzu `timestamp_t` to `gs::common::timestamp_t` to avoid conflict with neug (#827)
<!--
Thanks for your contribution! please review
https://github.com/GraphScope/neug/blob/main/CONTRIBUTING.md before
opening an issue.
-->

## What do these changes do?

This PR resolves a naming conflict between `timestamp_t` defined in
[Kuzu](https://github.com/GraphScope/neug/blob/include_rt_any/include/neug/compiler/common/types/timestamp_t.h#L32)
and `timestamp_t` defined in
[Neug](https://github.com/GraphScope/neug/blob/include_rt_any/include/neug/utils/property/types.h#L66).
In some cases within Kuzu, `timestamp_t` was passed to templated
functions without an explicit namespace, which caused the compiler to
incorrectly resolve it to `neug::timestamp_t` instead of the intended
one.

To fix this, this PR explicitly qualifies all template parameters
involving `timestamp_t` with the appropriate namespace `gs::common::`.
This ensures that the correct type is used and prevents ambiguous or
unintended resolutions during compilation.

<!-- Please give a short brief about these changes. -->

## Related issue number

<!-- Are there any issues opened that will be resolved by merging this
change? -->

Fixes