feat: Introduce `DeleteEdge` and `AddEdge` for EdgeTable (#1091)
Fix #1117

Implemented edge addition, deletion, and update for CSR, EdgeTable,
PropertyGraph, and UpdateTransaction. `NbrIterator` is used as the
unique identifier for edges at the storage level, while `vid + offset`
is used at the CSR level.

A challenge with deleting edges using `EdgeRecord` is the necessity to
scan for reverse-direction edges in other CSRs. We implemented a scan
function and propose using `src_vid + dst_vid + timestamp + data_value`
as a unique identifier for edges to be deleted in the other CSRs.